+++ test/test_algorithm.jl
@@ -150,7 +150,7 @@ end
                            "log_level" => 100)
 
     m = circle(solver=test_solver)
-    optimize!(m)
+    JuMP.optimize!(m)
 
     @test isapprox(objective_value(m), 1.4142135534556992; atol=1e-3)
 end
@@ -166,7 +166,7 @@ end
                            "log_level" => 100)
 
     m = circleN(solver=test_solver, N=4)
-    optimize!(m)
+    JuMP.optimize!(m)
     @test isapprox(objective_value(m), 2.0; atol=1e-3)
 end
 
@@ -300,7 +300,7 @@ end
                            "log_level" => 100)
 
     m = nlp3(solver=test_solver)
-    optimize!(m)
+    JuMP.optimize!(m)
 
     @test MOI.get(m, Alpine.NumberOfIterations()) == 1
     @test MOI.get(m, MOI.RawParameter("disc_ratio")) == 18
@@ -317,7 +317,7 @@ end
                            "log_level" => 100)
 
     m = nlp3(solver=test_solver)
-    optimize!(m)
+    JuMP.optimize!(m)
 
     @test MOI.get(m, Alpine.NumberOfIterations()) == 1
     @test MOI.get(m, MOI.RawParameter("disc_ratio")) == 14
@@ -334,7 +334,7 @@ end
                            "log_level" => 100)
 
     m = castro2m2(solver=test_solver)
-    optimize!(m)
+    JuMP.optimize!(m)
 
     @test MOI.get(m, Alpine.NumberOfIterations()) == 1
     @test MOI.get(m, MOI.RawParameter("disc_ratio")) == 8
@@ -351,7 +351,7 @@ end
                            "log_level" => 100)
 
     m = multi3N(solver=test_solver, N=3, exprmode=1)
-    optimize!(m)
+    JuMP.optimize!(m)
 
     @test MOI.get(m, Alpine.NumberOfIterations()) == 1
     @test MOI.get(m, MOI.RawParameter("disc_ratio")) == 16
@@ -368,7 +368,7 @@ end
                            "log_level" => 100)
 
     m = multi3N(solver=test_solver, N=3, exprmode=1)
-    optimize!(m)
+    JuMP.optimize!(m)
 
     @test MOI.get(m, Alpine.NumberOfIterations()) == 1
     @test MOI.get(m, MOI.RawParameter("disc_ratio")) == 20
@@ -385,7 +385,7 @@ end
                            "log_level" => 100)
 
     m = multi4N(solver=test_solver, N=2, exprmode=1)
-    optimize!(m)
+    JuMP.optimize!(m)
 
     @test MOI.get(m, Alpine.NumberOfIterations()) == 1
     @test MOI.get(m, MOI.RawParameter("disc_ratio")) == 12
@@ -402,7 +402,7 @@ end
                            "log_level" => 100)
 
     m = multi4N(solver=test_solver, N=2, exprmode=1)
-    optimize!(m)
+    JuMP.optimize!(m)
 
     @test MOI.get(m, Alpine.NumberOfIterations()) == 1
     @test MOI.get(m, MOI.RawParameter("disc_ratio")) == 20
@@ -419,7 +419,7 @@ end
                            "log_level" => 100)
 
     m = multi4N(solver=test_solver, N=2, exprmode=2)
-    optimize!(m)
+    JuMP.optimize!(m)
 
     @test MOI.get(m, Alpine.NumberOfIterations()) == 1
     @test MOI.get(m, MOI.RawParameter("disc_ratio")) == 20
@@ -433,7 +433,7 @@ end
                           "log_level" => 100)
 
     m = bpml_lnl(test_solver)
-    optimize!(m)
+    JuMP.optimize!(m)
     @test isapprox(objective_value(m), 0.3; atol=1e-6)
     alpine = JuMP.backend(m).optimizer.model
     @test haskey(alpine.nonconvex_terms, Expr[:(x[1]), :(x[6])])
@@ -457,7 +457,7 @@ end
                           "log_level" => 100)
 
     m = bpml_binl(test_solver)
-    optimize!(m)
+    JuMP.optimize!(m)
     @test isapprox(objective_value(m), 15422.058099086951; atol=1e-1)
 
     alpine = JuMP.backend(m).optimizer.model
@@ -517,7 +517,7 @@ end
 #                            "log_level" => 100)
 
 #     m = circle(solver=test_solver)
-#     optimize!(m)
+#     JuMP.optimize!(m)
 #     @test isapprox(objective_value(m), 1.4142135534556992; atol=1e-3)
 # end
 
@@ -572,7 +572,7 @@ end
                            "log_level" => 100)
 
     m = circle(solver=test_solver)
-    optimize!(m)
+    JuMP.optimize!(m)
     @test isapprox(objective_value(m), 1.4142135534556992; atol=1e-3)
 end
 
@@ -664,7 +664,7 @@ end
                            "time_limit" => 100000,
                            "log_level" => 100)
     m = nlp3(solver=test_solver)
-    optimize!(m)
+    JuMP.optimize!(m)
     @test isapprox(objective_bound(m), 6561.7841; atol=1e-3)
 end